pyproject.toml
[tool.poetry]
name = "market_bot"
version = "0.1.0"
description = ""
authors = ["Vasker <v92prog@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.13"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
